fixes line item delete

+++ common/models/InvoiceLineItem.php
@@ -285,13 +285,15 @@ class InvoiceLineItem extends \yii\db\ActiveRecord
     public function beforeSoftDelete()
     {
         if ($this->invoice->isInvoice() && $this->isLessonItem()) {
-            if ($this->lesson->hasCreditUsed($this->enrolment->id)) {
-                $lessonCreditPayments = $this->lesson->getCreditUsedPayment($this->enrolment->id);
-                foreach ($lessonCreditPayments as $lessonCreditPayment) {
-                    $lessonCreditPayment->delete();
+            if ($this->lesson->isPrivate()) {
+                if ($this->lesson->hasCreditUsed($this->enrolment->id)) {
+                    $lessonCreditPayments = $this->lesson->getCreditUsedPayment($this->enrolment->id);
+                    foreach ($lessonCreditPayments as $lessonCreditPayment) {
+                        $lessonCreditPayment->delete();
+                    }
                 }
+                $this->lesson->unschedule();
             }
-            $this->lesson->unschedule();
         }
 
         return true;
